pyproject.toml
[project]
name = "ifwebuildit"
version = "0.1.0"
description = "Knowledge Base Overnight Enrichment — Azure AI Content Understanding demo"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12,<3.13"
dependencies = [
    # Azure AI Content Understanding
    "azure-ai-contentunderstanding>=1.0.0b1",

    # Azure AI Search (preview for agentic retrieval)
    "azure-search-documents==11.7.0b2",

    # Azure Storage
    "azure-storage-blob",
    "azure-identity>=1.25.1",

    # Azure OpenAI (embeddings + chat completions)
    "openai>=1.0",

    # HTTP/API
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv",
    "python-multipart",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ifwebuildit"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "pyright",
    "pre-commit",
    "detect-secrets",
]

[tool.uv]
prerelease = "allow"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=80",
]

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
include = ["src"]
exclude = ["tests/fixtures", ".venv"]
